+++ scripts/macros/2024/spells/gaseousForm.js
@@ -1,0 +1,6 @@
+import {gaseousForm as gaseousFormLegacy} from '../../../legacyMacros.js';
+export let gaseousForm = {
+    ...gaseousFormLegacy,
+    version: '1.2.24',
+    rules: 'modern'
+};